Cleaned up javadoc and BiPredicate<BloomFilter,BloomFilter> processing
